moved dto, entity packages to new package: dao added ForecastRepository and TimeIntervalRepository added 'hibernate.default_schema' to application.yaml and application-auto-test.yaml fixed 'preConditions' of main changelog fixed @ToString and @EqualsAndHashCode to ForecastEntity
