+++ src/test/resources/application-auto-test.yaml
@@ -9,8 +9,10 @@ spring:
     open-in-view: false
     hibernate:
       ddl-auto: none
+    properties:
+      hibernate.default_schema: geomagnetic
   datasource:
-    url: jdbc:h2:mem:geomagnetic;MODE=PostgreSQL;INIT=CREATE SCHEMA IF NOT EXISTS geomagnetic;DATABASE_TO_UPPER=false
+    url: jdbc:h2:mem:geomagnetic;MODE=PostgreSQL;INIT=CREATE SCHEMA IF NOT EXISTS geomagnetic;DATABASE_TO_LOWER=true
     username: username
     password: password
     driver-class-name: org.h2.Driver
@@ -24,6 +26,7 @@ spring:
       minimum-idle: 2
       schema: geomagnetic
   liquibase:
+    enabled: true
     default-schema: geomagnetic
     user: username
     password: password
